+++ programmable_video/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.13.0
 
+- **Android**: Fixed compilation issue for Flutter > 2.12
 - **BREAKING**: Updated permission_handler dependency from 8.3.0 to 9.2.0
 - Added a `create` function to the `LocalVideoTrack` class that will trigger initialization at the native layer.
 - Maintain a map of `LocalVideoTracks` at the native layer to avoid initializing a second track with the same id should the end developer then provide this when connecting.